# docs/api/swagger.yaml
basePath: /api
definitions:
...
  internal_modules_user.CreateRequest:
    properties:
      avatar_id:
        type: integer
      gender:
        enum:
        - 0
        - 1
        - 2
        type: integer
      mobile:
        type: string
      nick_name:
        maxLength: 64
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      status:
        enum:
        - 1
        - 2
        type: integer
      store_id:
        type: integer
      username:
        maxLength: 32
        minLength: 5
        type: string
    required:
    - gender
    - mobile
    - nick_name
    - password
    - status
    - username
    type: object
  internal_modules_user.DeleteRequest:
    properties:
      id:
        type: integer
      store_id:
        type: integer
    required:
    - id
    type: object
  internal_modules_user.DetailRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  internal_modules_user.ListRequest:
    properties:
      limit:
        maximum: 100
        minimum: 1
        type: integer
      mobile:
        type: string
      page:
        minimum: 1
        type: integer
      status:
        type: integer
      store_id:
        type: integer
      username:
        type: string
    required:
    - limit
    - page
    type: object
  internal_modules_user.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
  internal_modules_user.PointsChangeRequest:
    properties:
      change_type:
        enum:
        - 1
        - 2
        type: integer
      points:
        type: integer
      remark:
        maxLength: 255
        type: string
      source_id:
        maxLength: 50
        type: string
      source_type:
        type: integer
      store_id:
        type: integer
      user_id:
        type: integer
    required:
    - change_type
    - points
    - user_id
    type: object
  internal_modules_user.PointsLogListRequest:
    properties:
      change_type:
        enum:
        - 0
        - 1
        - 2
        type: integer
      end_date:
        type: string
      limit:
        maximum: 100
        minimum: 1
        type: integer
      page:
        minimum: 1
        type: integer
      start_date:
        type: string
      store_id:
        type: integer
      user_id:
        type: integer
    required:
    - limit
    - page
    - user_id
    type: object
  internal_modules_user.UpdateRequest:
    properties:
      avatar_id:
        type: integer
      gender:
        enum:
        - 0
        - 1
        - 2
        type: integer
      id:
        type: integer
      mobile:
        type: string
      nick_name:
        maxLength: 64
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      status:
        enum:
        - 1
        - 2
        type: integer
      store_id:
        type: integer
      username:
        maxLength: 32
        minLength: 5
        type: string
    required:
    - gender
    - id
    - mobile
    - nick_name
    - status
    - username
    type: object
  internal_modules_user.UploadFile:
    properties:
      domain:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      id:
        type: integer
    type: object
  internal_modules_user.User:
    properties:
      avatar:
        $ref: '#/definitions/internal_modules_user.UploadFile'
      avatar_id:
        type: integer
      city:
        type: string
      country:
        type: string
      created_at:
        type: integer
      gender:
        type: integer
      id:
        type: integer
      last_login_time:
        type: integer
      mobile:
        type: string
      nick_name:
        type: string
      platform:
        type: string
      points:
        type: integer
      province:
        type: string
      status:
        type: integer
      store_id:
        type: integer
      updated_at:
        type: integer
      username:
        type: string
    type: object
  internal_modules_user.UserLoginResponse:
    properties:
      token:
        type: string
      ttl:
        type: integer
      user:
        $ref: '#/definitions/internal_modules_user.User'
    type: object
  response.Response:
    properties:
      cost:
        type: string
      data: {}
      errcode:
        type: integer
      message:
        type: string
      traceId:
        type: string
    type: object
  zero-backend_internal_modules_rbac.AuthLoginRequest:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha_code
    - captcha_id
    - password
    - username
    type: object
  zero-backend_internal_modules_rbac.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 20
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
host: localhost:8081
info:
  contact: {}
  title: Zero API
  version: "1.0"
paths:
...
  /change-password:
    post:
      parameters:
      - description: 修改密码参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/zero-backend_internal_modules_rbac.ChangePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户修改密码
      tags:
      - 用户认证
...
  /login:
    post:
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/zero-backend_internal_modules_rbac.AuthLoginRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_user.UserLoginResponse'
              type: object
      summary: 用户登录
      tags:
      - 用户认证
...
  /user/points/change:
    post:
      parameters:
      - description: 变更参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.PointsChangeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户积分变更
      tags:
      - 用户管理
  /user/points/logs:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.PointsLogListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_user.ListResult'
              type: object
      summary: 获取用户积分记录
      tags:
      - 用户管理
  /user/user/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.CreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建用户
      tags:
      - 用户管理
  /user/user/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.DeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除用户
      tags:
      - 用户管理
  /user/user/detail:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.DetailRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取用户详情
      tags:
      - 用户管理
  /user/user/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.ListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_user.ListResult'
              type: object
      summary: 获取用户列表
      tags:
      - 用户管理
  /user/user/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.UpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新用户
      tags:
      - 用户管理
swagger: "2.0"